+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 ## Added
 
 - **Drainer** - A special task-agent in the superagent crate to drain events sources.
+- **ToRecipient** - A trait to produce a recipient for events.
 
 ## Improved